requirements.txt
python3-discogs-client~=2.8
python-dotenv~1.1.0